README.md Added
A simple program to check whether the domain is alive or not basically. This script can be made advanced by checking the list of all subdomains for a domain is alive or not. 
I am learning concurrency in Go but once I finish,I will definitely update it so that you can take full advantage of it.